+++ ea_airflow_util/callables/sharefile.py
@@ -1,8 +1,10 @@
 import logging
 import os
+import pathlib
 import pytz
 import re
 import requests
+import uuid
 
 from datetime import datetime
 from typing import List, Optional
@@ -10,6 +12,7 @@ from typing import List, Optional
 from airflow.exceptions import AirflowException, AirflowSkipException, AirflowFailException
 
 from ea_airflow_util.providers.sharefile.hooks.sharefile import SharefileHook
+from airflow.providers.amazon.aws.hooks.s3 import S3Hook
 
 def list_sharefile_objects(sharefile_conn_id: str, remote_dir: str) -> List[str]:
     sharefile_hook = SharefileHook(sharefile_conn_id)
@@ -118,6 +121,38 @@ def sharefile_to_disk(
 
     return local_path
 
+
+
+def disk_to_sharefile(sf_conn_id: str, sf_folder_path: str, local_path: str):
+    """Post a file or the contents of a directory to the specified Sharefile folder"""
+    sf_hook = SharefileHook(sf_conn_id )
+
+    sf_folder_id = sf_hook.folder_id_from_path(sf_folder_path)
+    if sf_folder_id is None:
+        raise AirflowException(f"failed to find Sharefile folder {sf_folder_path}")
+
+    local_path = pathlib.Path(local_path)
+    if local_path.is_dir():
+        # upload all files in directory
+        for filepath in local_path.iterdir():
+            sf_hook.upload_file(sf_folder_id, filepath)
+    else:
+        sf_hook.upload_file(sf_folder_id, local_path)
+
+
+def s3_to_sharefile(s3_conn_id: str, s3_key: str, sf_conn_id: str, sf_folder_path: str):
+    """Copy a single file from S3 to Sharefile"""
+
+    s3_hook = S3Hook(s3_conn_id)
+    s3_creds = s3_hook.get_connection(s3_hook.aws_conn_id)
+    s3_bucket = s3_creds.schema
+
+    downloaded_file = s3_hook.download_file(s3_key, s3_bucket, preserve_file_name=True)
+
+    disk_to_sharefile(sf_conn_id, sf_folder_path, downloaded_file)
+
+    # claen up the disk
+    os.remove(downloaded_file)
 
 def check_for_new_files(sharefile_conn_id: str, sharefile_path: str, num_expected_files: Optional[int] = None, updated_after: Optional[datetime] = None):
     """