+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # ea_airflow_util v0.3.2
 ## New features
 - `AWSParamStoreToAirflowDAG` allows more flexibility when passing Parameter Store paths. Use `{tenant_code}` when the tenant is in the middle of the path, instead of the end.
+- Add `s3_to_sharefile` and `disk_to_sharefile` callables
+- Add methods to the `SharefileHook`
 - Add ShareFile callable `check_for_new_files()` to assert expectations in ShareFile directory
 - Add `S3ToSnowflakeOperator` to S3 operators
 - Add `LoadSharefileCustomUsersDag` to automate Heimdall user creation from uploaded authenticated users files in ShareFile